Template should not impose ``title`` formatting
If the teaser title is entered as "IBM News", the default template should not forcefully reformat this as "Ibm News". If users are comfortable using the ``title`` filter as a site-wide policy, they can easily override the template. But in the default case, it should assume the content editors are consenting adults capable of correctly using upper and lower case letters.